// types.ts
export interface PromptElements {
  subject: string;
  mainAction: string;
  keyElements: string;
  visualStyle: string;
  artisticInfluence: string;
  cameraAngle: string;
  cameraMovement: string;
  lightingStyle: string;
  colorPalette: string;
  setting: string;
  timeOfDay: string;
  mood: string;
  aspectRatio: string;
  negativePrompt: string;
  additionalDetails: string;
}

export interface SelectOption {
  value: string;
  label: string;
}